// src/pages/annotator/tools/grabcut-tool/grabcut-tool.enums.tsx
// INTEL CONFIDENTIAL
//
// Copyright (C) 2021 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

export enum GrabcutToolType {
    InputTool = 'grabcut-input-tool',
    BackgroundTool = 'grabcut-background-tool',
    ForegroundTool = 'grabcut-foreground-tool',
}

export enum GrabcutToolLabel {
    InputTool = 'New selection',
    ForegroundTool = 'Mark object',
    BackgroundTool = 'Mark background',
}

export enum GrabcutToolIconsOffset {
    Foreground = '16 16',
    Background = '16 16',
}